+++ src/hera/template_ref.py
@@ -11,11 +11,16 @@ class TemplateRef(BaseModel):
         The name of the workflow template reference.
     template: str
         The name of the independent task template to reference in the workflow template.
+    cluster_scope: bool = False
+        Whether the referenced template is cluster scoped or not.
     """
 
     name: str
     template: str
+    cluster_scope: bool = False
 
     @property
     def argo_spec(self) -> IoArgoprojWorkflowV1alpha1TemplateRef:
-        return IoArgoprojWorkflowV1alpha1TemplateRef(name=self.name, template=self.template)
+        return IoArgoprojWorkflowV1alpha1TemplateRef(
+            name=self.name, template=self.template, cluster_scope=self.cluster_scope
+        )