+++ src/hera/workflow.py
@@ -4,6 +4,7 @@ from typing import Dict, List, Optional, Tuple
 from argo_workflows.models import (
     IoArgoprojWorkflowV1alpha1DAGTemplate,
     IoArgoprojWorkflowV1alpha1Template,
+    IoArgoprojWorkflowV1alpha1VolumeClaimGC,
     IoArgoprojWorkflowV1alpha1Workflow,
     IoArgoprojWorkflowV1alpha1WorkflowSpec,
     IoArgoprojWorkflowV1alpha1WorkflowTemplateRef,
@@ -11,9 +12,11 @@ from argo_workflows.models import (
     ObjectMeta,
 )
 
+from hera.host_alias import HostAlias
 from hera.security_context import WorkflowSecurityContext
 from hera.task import Task
 from hera.ttl_strategy import TTLStrategy
+from hera.volume_claim_gc import VolumeClaimGCStrategy
 from hera.workflow_editors import add_head, add_tail, add_task, add_tasks
 from hera.workflow_service import WorkflowService
 
@@ -52,6 +55,10 @@ class Workflow:
         If you create a WorkflowTemplate resource either clusterWorkflowTemplate or not (clusterScope attribute bool)
         you can reference it again and again when you create a new Workflow without specifying the same tasks and
         dependencies. Official doc: https://argoproj.github.io/argo-workflows/fields/#workflowtemplateref
+    volume_claim_gc_strategy: Optional[VolumeClaimGCStrategy] = None
+        Define how to delete volumes from completed Workflows.
+    host_aliases: Optional[List[HostAlias]] = None
+        Mappings between IP and hostnames.
     """
 
     def __init__(
@@ -67,6 +74,8 @@ class Workflow:
         image_pull_secrets: Optional[List[str]] = None,
         workflow_template_ref: Optional[str] = None,
         ttl_strategy: Optional[TTLStrategy] = None,
+        volume_claim_gc_strategy: Optional[VolumeClaimGCStrategy] = None,
+        host_aliases: Optional[List[HostAlias]] = None,
     ):
         self.name = f'{name.replace("_", "-")}'  # RFC1123
         self.namespace = namespace or 'default'
@@ -105,6 +114,16 @@ class Workflow:
 
         if ttl_strategy:
             setattr(self.spec, 'ttl_strategy', ttl_strategy.argo_ttl_strategy)
+
+        if volume_claim_gc_strategy:
+            setattr(
+                self.spec,
+                'volume_claim_gc',
+                IoArgoprojWorkflowV1alpha1VolumeClaimGC(strategy=volume_claim_gc_strategy.value),
+            )
+
+        if host_aliases:
+            setattr(self.spec, 'host_aliases', [h.argo_host_alias for h in host_aliases])
 
         if self.security_context:
             security_context = self.security_context.get_security_context()