+++ tests/test_workflow.py
@@ -2,11 +2,15 @@ from unittest.mock import Mock
 
 import pytest
 from argo_workflows.model.pod_security_context import PodSecurityContext
+from argo_workflows.models import HostAlias as ArgoHostAlias
 
+from hera.host_alias import HostAlias
 from hera.resources import Resources
 from hera.security_context import WorkflowSecurityContext
 from hera.task import Task
+from hera.template_ref import TemplateRef
 from hera.ttl_strategy import TTLStrategy
+from hera.volume_claim_gc import VolumeClaimGCStrategy
 from hera.volumes import (
     ConfigMapVolume,
     EmptyDirVolume,
@@ -235,3 +239,44 @@ def test_wf_adds_ttl_strategy(ws):
     }
 
     assert w.spec.ttl_strategy._data_store == expected_ttl_strategy
+
+
+def test_wf_adds_volume_claim_gc_strategy_on_workflow_completion(ws):
+    w = Workflow('w', service=ws, volume_claim_gc_strategy=VolumeClaimGCStrategy.OnWorkflowCompletion)
+
+    expected_volume_claim_gc = {"strategy": "OnWorkflowCompletion"}
+
+    assert w.spec.volume_claim_gc._data_store == expected_volume_claim_gc
+
+
+def test_wf_adds_volume_claim_gc_strategy_on_workflow_success(ws):
+    w = Workflow('w', service=ws, volume_claim_gc_strategy=VolumeClaimGCStrategy.OnWorkflowSuccess)
+
+    expected_volume_claim_gc = {"strategy": "OnWorkflowSuccess"}
+
+    assert w.spec.volume_claim_gc._data_store == expected_volume_claim_gc
+
+
+def test_wf_adds_host_aliases(ws):
+    w = Workflow(
+        'w',
+        service=ws,
+        host_aliases=[
+            HostAlias(hostnames=["host1", "host2"], ip="0.0.0.0"),
+            HostAlias(hostnames=["host3"], ip="1.1.1.1"),
+        ],
+    )
+
+    assert w.spec.host_aliases[0] == ArgoHostAlias(hostnames=["host1", "host2"], ip="0.0.0.0")
+    assert w.spec.host_aliases[1] == ArgoHostAlias(hostnames=["host3"], ip="1.1.1.1")
+
+
+def test_wf_add_task_with_template_ref(w):
+    t = Task("t", template_ref=TemplateRef(name="name", template="template"))
+    w.add_task(t)
+
+    assert w.dag_template.tasks[0] == t.argo_task
+
+    # Not add a Task with TemplateRef to w.spec.templates
+    # Note: w.spec.templates[0] is a template of dag
+    assert len(w.spec.templates) == 1